requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.6b0
click==8.0.1
Django==3.2.5
environ==1.0
environs==9.3.2
flake8==3.9.2
marshmallow==3.12.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
django-multi-form-view
